Add permission to board|pedal parameters in <board|pedal>_params
Users now have access to these parameters and can set them, but for
some reason they are saving as strings and not decimals despite my
schema saying otherwise (open issue on tracker).

Also added validation for length and width on pedal so you cannot
make a pedal with null dimensions. probably should have a minimum of
1 inch or something.